02/05

+++ src/main/java/com/projet/tp/dao/Type_prodRepo.java
@@ -5,6 +5,6 @@ package com.projet.tp.dao;
 import com.projet.tp.entity.Type_produit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Type_prodRepo extends JpaRepository<Type_produit, Long> {
-    //public Type_produit findByName(String nomtype);
+public interface Type_prodRepo extends JpaRepository<Type_produit, Integer> {
+
 }
